Break the library package apart
Also:
* Fix up Resolvr semantics
