+++ service/library/library.go
@@ -3,22 +3,9 @@ package library
 
 import (
 	"fmt"
-	"os"
-	"runtime"
-	"sort"
-
-	"github.com/fubarhouse/pygmy/v1/service/amazee"
-	"github.com/fubarhouse/pygmy/v1/service/dnsmasq"
-	"github.com/fubarhouse/pygmy/v1/service/haproxy"
-	"github.com/fubarhouse/pygmy/v1/service/haproxy_connector"
 	model "github.com/fubarhouse/pygmy/v1/service/interface"
-	"github.com/fubarhouse/pygmy/v1/service/mailhog"
-	"github.com/fubarhouse/pygmy/v1/service/network"
 	"github.com/fubarhouse/pygmy/v1/service/resolv"
-	"github.com/fubarhouse/pygmy/v1/service/ssh/agent"
-	"github.com/fubarhouse/pygmy/v1/service/ssh/key"
 	"github.com/imdario/mergo"
-	"github.com/spf13/viper"
 )
 
 // Config is a struct of configurable options which can
@@ -62,249 +49,4 @@ func mergeService(destination model.Service, src *model.Service) (*model.Service
 func getService(s model.Service, c model.Service) model.Service {
 	Service, _ := mergeService(s, &c)
 	return *Service
-}
-
-func SshKeyAdd(c Config, key string) {
-
-	if c.SkipKey {
-		return
-	}
-
-	Setup(&c)
-
-	if key != "" {
-		if _, err := os.Stat(key); err != nil {
-			fmt.Printf("The file path %v does not exist, or is not readable.\n%v\n", key, err)
-			return
-		}
-	}
-
-	if !agent.Search(key) {
-
-		if key != "" {
-			c.Key = key
-		}
-
-		//data, _ := model.Start(Service)
-		//fmt.Println(string(data))
-
-	} else {
-		fmt.Printf("Already added key file %v.\n", c.Key)
-	}
-}
-
-func Clean(c Config) {
-
-	Setup(&c)
-	Containers, _ := model.DockerContainerList()
-
-	for _, Container := range Containers {
-		if l := Container.Labels["pygmy"]; l == "pygmy" {
-
-			err := model.DockerKill(Container.ID)
-			if err == nil {
-				fmt.Printf("Successfully killed  %v.\n", Container.Names[0])
-			}
-
-			err = model.DockerRemove(Container.ID)
-			if err == nil {
-				fmt.Printf("Successfully removed %v.\n", Container.Names[0])
-			}
-		}
-	}
-
-	for _, resolver := range c.Resolvers {
-		resolv.New(resolv.Resolv{Name: string(resolver.Name), Data: string(resolver.Data), Path: string(resolver.Path)}).Clean()
-	}
-}
-
-func Restart(c Config) {
-	Down(c)
-	Up(c)
-}
-
-func Status(c Config) {
-
-	Setup(&c)
-
-	for Label, Service := range c.Services {
-		if !Service.Disabled && !Service.Discrete {
-			if s, _ := Service.Status(); s {
-				fmt.Printf("[*] %v: Running as container %v\n", Label, Service.Name)
-			} else {
-				fmt.Printf("[ ] %v is not running\n", Label)
-			}
-		}
-	}
-
-	for Network, Containers := range c.Networks {
-		netStat, _ := network.Status(Network)
-		if netStat {
-			for _, Container := range Containers {
-				if s, _ := haproxy_connector.Connected(Container, Network); s {
-					fmt.Printf("[*] %v is connected to network %v\n", Container, Network)
-				} else {
-					fmt.Printf("[ ] %v is not connected to network %v\n", Container, Network)
-				}
-			}
-		}
-	}
-
-	for _, resolver := range c.Resolvers {
-		r := resolv.Resolv{Name: resolver.Name, Path: resolver.Path, Data: resolver.Data}
-		if s := r.Status(); s {
-			fmt.Printf("[*] Resolv %v is properly conneted\n", resolver.Name)
-		} else {
-			fmt.Printf("[ ] Resolv %v is not properly connected\n", resolver.Name)
-		}
-	}
-
-	for _, volume := range c.Volumes {
-		if s, _ := model.DockerVolumeExists(volume); s {
-			fmt.Printf("[*] Volume %v has been created\n", volume)
-		} else {
-			fmt.Printf("[ ] Volume %v has not been created\n", volume)
-		}
-	}
-
-}
-
-func Down(c Config) {
-
-	Setup(&c)
-
-	for _, Service := range c.Services {
-		if !Service.Disabled {
-			Service.Stop()
-		}
-	}
-
-	if !c.SkipResolver {
-		for _, resolver := range c.Resolvers {
-			resolv.New(resolv.Resolv{Name: string(resolver.Name), Data: string(resolver.Data), Path: string(resolver.Path)}).Clean()
-		}
-	}
-}
-
-func Setup(c *Config) {
-
-	viper.SetDefault("Networks", map[string][]string{
-		"amazeeio-network": []string{
-			"amazeeio-haproxy",
-		},
-	})
-
-	var ResolvOSX = resolv.Resolv{
-		Data: "\n# Generated by amazeeio pygmy\nnameserver 127.0.0.1\nport 6053\n",
-		File: "docker.amazee.io",
-		Path: "/etc/resolver",
-	}
-
-	var ResolvGeneric = resolv.Resolv{
-		File: "resolv.conf",
-		Data: "\nnameserver 127.0.0.1 # added by amazee.io pygmy",
-		Path: "/etc",
-	}
-
-	DefaultResolvers := []resolv.Resolv{}
-
-	if runtime.GOOS == "darwin" {
-		fmt.Println(ResolvOSX)
-		DefaultResolvers = append(DefaultResolvers, ResolvOSX)
-	} else if runtime.GOOS == "linux" {
-		DefaultResolvers = append(DefaultResolvers, ResolvGeneric)
-	}
-
-	viper.SetDefault("Resolvers", &DefaultResolvers)
-	fmt.Println(viper.GetString("Resolvers"))
-
-	e := viper.Unmarshal(&c)
-
-	if e != nil {
-		fmt.Println(e)
-	}
-
-	// If Services have been provided in complete or partially,
-	// this will override the defaults allowing any value to
-	// be changed by the user in the configuration file ~/.pygmy.yml
-	c.Services["amazeeio-ssh-agent-show-keys"] = getService(key.NewShower(), c.Services["amazeeio-ssh-agent-show-keys"])
-	c.Services["amazeeio-ssh-agent-add-key"] = getService(key.NewAdder(c.Key), c.Services["amazeeio-ssh-agent-add-key"])
-	c.Services["amazeeio-dnsmasq"] = getService(dnsmasq.New(), c.Services["amazeeio-dnsmasq"])
-	c.Services["amazeeio-haproxy"] = getService(haproxy.New(), c.Services["amazeeio-haproxy"])
-	c.Services["mailhog.docker.amazee.io"] = getService(mailhog.New(), c.Services["mailhog.docker.amazee.io"])
-	c.Services["amazeeio-ssh-agent"] = getService(agent.New(), c.Services["amazeeio-ssh-agent"])
-	c.SortedServices = make([]string, 0, len(c.Services))
-
-	// We need services to be sortable...
-	for key := range c.Services {
-		c.SortedServices = append(c.SortedServices, key)
-	}
-	sort.Strings(c.SortedServices)
-
-}
-
-func Up(c Config) {
-
-	Setup(&c)
-
-	for _, volume := range c.Volumes {
-		if s, _ := model.DockerVolumeExists(volume); !s {
-			_, err := model.DockerVolumeCreate(volume)
-			if err != nil {
-				fmt.Println(err)
-			}
-		} else {
-			fmt.Printf("Already created volume %v\n", volume)
-		}
-	}
-
-	// Maps are... bad for predictable sequencing.
-	// Look over the sorted slice and start them in
-	// alphabetical order - so that one can configure
-	// an ssh-agent like amazeeio-ssh-agent.
-	for _, service := range c.SortedServices {
-		s := c.Services[service]
-		if !s.Disabled {
-			s.Start()
-		}
-	}
-
-	for Network, Containers := range c.Networks {
-		netStat, _ := network.Status(Network)
-		if !netStat {
-			network.Create(Network)
-		}
-		for _, Container := range Containers {
-			if s, _ := haproxy_connector.Connected(Container, Network); !s {
-				haproxy_connector.Connect(Container, Network)
-				if s, _ := haproxy_connector.Connected(Container, Network); s {
-					fmt.Printf("Successfully connected %v to %v\n", Container, Network)
-				} else {
-					fmt.Printf("Could not connect %v to %v\n", Container, Network)
-				}
-			} else {
-				fmt.Printf("Already connected %v to %v\n", Container, Network)
-			}
-		}
-	}
-
-	if !c.SkipResolver {
-		for _, resolver := range c.Resolvers {
-			resolv.New(resolv.Resolv{Name: string(resolver.Name), Data: string(resolver.Data), Path: string(resolver.Path)}).Configure()
-		}
-	}
-
-	if !c.SkipKey {
-
-		SshKeyAdd(c, c.Key)
-
-	}
-}
-
-func Update(c Config) {
-	amazee.AmazeeImagePull()
-}
-
-func Version(c Config) {
-	fmt.Println("version called")
 }
