+++ service/library/restart.go
@@ -1,0 +1,6 @@
+package library
+
+func Restart(c Config) {
+	Down(c)
+	Up(c)
+}
