+++ service/library/update.go
@@ -1,0 +1,7 @@
+package library
+
+import "github.com/fubarhouse/pygmy/v1/service/amazee"
+
+func Update(c Config) {
+	amazee.AmazeeImagePull()
+}
