+++ service/library/version.go
@@ -1,0 +1,7 @@
+package library
+
+import "fmt"
+
+func Version(c Config) {
+	fmt.Println("version called")
+}
